Renaming solnpool.py to gurobi_solnpool.py
