+++ pyomo/contrib/alternative_solutions/__init__.py
@@ -11,7 +11,7 @@
 
 from pyomo.contrib.alternative_solutions.aos_utils import logcontext
 from pyomo.contrib.alternative_solutions.solution import Solution
-from pyomo.contrib.alternative_solutions.solnpool import gurobi_generate_solutions
+from pyomo.contrib.alternative_solutions.gurobi_solnpool import gurobi_generate_solutions
 from pyomo.contrib.alternative_solutions.balas import enumerate_binary_solutions
 from pyomo.contrib.alternative_solutions.obbt import (
     obbt_analysis,
