+++ pyomo/contrib/alternative_solutions/gurobi_solnpool.py
